refactor: adjust some minor things
