--- tools/fill-epic-store-ids/.gitignore
@@ -1,1 +1,0 @@
-*.js
