--- tools/fill-epic-store-ids/.prettierrc.json
@@ -1,1 +1,0 @@
-{}
