--- tools/fill-epic-store-ids/interfaces.ts
@@ -1,2 +1,0 @@
-export type EpicDataOffersLists = (string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number | string | (string)[] | null | number)[] | null;
-
