--- tools/fill-epic-store-ids/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "inlineSourceMap": true,
-    "allowSyntheticDefaultImports": true
-  },
-  "include": ["index.ts"]
-}
